Add measurement files derived from pattern files
Add an initial set of measurement files, with the measurements
taken from the patternshare pattern files.

In order to make any pattern work with any set of measurements,
each set of measurements has been converted to a superset of all
measurements appearing in all patterns. This is consistent with
eventually moving to a uniformly defined set of measurements.

Signed-off-by: Steve Conklin <sconklin@canonical.com>